requirements.txt
django
djangorestframework
environs
psycopg2-binary
python-dotenv
